common_requirements.txt
python-dotenv >= 1.0.1, < 2
langchain-core == 0.2.27
langchain-community == 0.2.11
langchain-openai == 0.1.20
langchainhub == 0.1.21
langsmith == 0.1.99
ragstack-ai-knowledge-store >= 0.2.1, < 0.3
simsimd >= 5.0.0, < 6.0.0
tqdm >= 4.66.4, < 4.67